fix: response validation with kin-openapi (#5)
Response format validation was relying on `client.ParseGetDriverJourneysResponse`, which contrarily to what was expected, does not validate the response (e.g. missing required fields are set to default zero value).

Use of [getkin/kin-openapi](https://github.com/getkin/kin-openapi) for response validation instead. 

Some refactoring was made, in particular, introduction of a type `CriticAssertion`, and of method `(AssertionAccumulator) Run(...Assertion)`. A failure of a `CriticAssertion` stops the subsequent executions of next assertions in a run.