refactoring: endpoint dependant implementations in their own file (#11)
This refactoring puts endpoint dependent code into their own files (`test_exported.go` and `test_implementation.go`).

This makes it easier to add new tests, with `test_exported.go` as an entry point. 

In addition, the tests functions to be run on a specific endpoint can be registered, rather than hardcoded. This gives more flexibility for adapting the test suite.