feat: Generate test data and test commands from golang tests (#40)
The objective of this PR is to be able to generate golang tests of the
example API into two separate files :
- one with all data for the tests. Note that all data of all tests are
put in a single file (which could lead to some conflicts between the
tests).
- another one with the commands that reproduce the tests. 


To generate the files, go in "./cmd/service" and enter `go generate` (or
`go test -generate`). The files are generated in "./cmd/service/data"
directory.

Only GET/PATCH/POST /bookings, POST /booking_events, POST /messages are
currently supported, as it is more tricky for the search endpoints not
to collide.

Among the features brought by this PR: 

* The new test files are tested in CI (see changes on
".github/workflows/go_tests_lint.yml")
* The "serve" command nows accept a `--data` flag that loads custom data
by providing a path to a JSON file.

Among the fixes brought by this PR:

* the bare `test` command was outdated, with missing handling of the
request method, and of request bodies. This is now fixed with the flag
`--method` (GET by default) and by reading the body from StdIn (if any),
like the other test commands.
* The PatchBooking test was not testing booking status correctly. Now
fixed.

Among the refactorings brought by this PR:

* the API test helpers ("cmd/service/api_test.go") are factorize with an
`APITestHelper` interface (finally !)
* To avoid collisions, PostMessage test now use different users for each
test.
* the API tests are now run in named subtests.