feat: generated example test commands for GET /driver_journeys and GET /passenger_journeys (#51)
This PR improve test generations, example test commands for GET
/driver_journeys and GET /passenger_journeys are now generated as well.

To isolate (as much as possible) the tests without isolating the data,
each test data is set to a different week. In this way, as soon as the
query filtering by date is working well, each command is rightly
associated with its data.

This is as much as can be done if we do not want to load the data for
each test separately.